+++ package.json
@@ -15,7 +15,8 @@
     "d3": "^7.4.2",
     "danfojs": "^1.1.1",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "react-router-dom": "6"
   },
   "devDependencies": {
     "@types/d3": "^7.1.0",
